+++ publiccms-by-gradle/src/com/publiccms/views/directive/cms/CmsLotteryUserListDirective.java
@@ -17,9 +17,9 @@ public class CmsLotteryUserListDirective extends AbstractTemplateDirective {
 
     @Override
     public void execute(RenderHandler handler) throws IOException, Exception {
-        PageHandler page = service.getPage(handler.getInteger("lotteryId"), handler.getLong("userId"),
-                handler.getBoolean("winning"), handler.getDate("startCreateDate"), handler.getDate("endCreateDate"),
-                handler.getString("orderType"), handler.getInteger("pageIndex", 1), handler.getInteger("count", 30));
+        PageHandler page = service.getPage(handler.getLong("lotteryId"), handler.getLong("userId"), handler.getBoolean("winning"),
+                handler.getDate("startCreateDate"), handler.getDate("endCreateDate"), handler.getString("orderType"),
+                handler.getInteger("pageIndex", 1), handler.getInteger("count", 30));
         handler.put("page", page).render();
     }
